Updated Blackwall.js
Line 46: Fixed an issue with a copy of policy being stored in Bloc
instance.
Line 67: Updated Session creation function to be sync. (any I/O in the
future is probably implemented within a policy and since sessions are
now assigned through the assign function this function can be left as a
sync function)
Line 71: New Function that assigns a policy to a session