Bloc Changes
Line 18: If Member is not found the storage is set to a new object.
Line 39: [Feature] Add Option to move attached Data to FS
Line 47: Added Name to Function
Line 67: Added Clean Function that cleans the Bloc Members' Keys which
have already been set to undefined. (this function is expensive)
Line 75: Added Name to Function