Make Bloc a part of Blackwall instance
Blocs are no longer specific to a Policy. This allows for better
clustering as a Bloc's Members could be fully managed by a database such
as Redis or Memcached while the policies and sessions remain
independent. Blocs will still be assigned to policies to store and
handle sessions. This also makes policies independent of the clustering
process and thus one policy can be applied to multiple clusters or
instances.